Fix: small changes like putting the logout trigger in the headerpopupaccount.vue instead of the landingspage.vue and making that button work to go to the materiaalpagina. Fix: not every url is blacklisted if you aren't logged in
